Improved ALPS support.
Update issue 17

To date ALPS is not enabled by default and under active development.

+++ kernel/random.h
@@ -94,7 +94,7 @@ namespace vita
   const T &random::element(const std::vector<T> &vect)
   {
     assert(vect.size());
-    return vect[between<unsigned>(0, vect.size())];
+    return vect[between<size_t>(0, vect.size())];
   }
 
   ///
